requirements.txt
numpy>=1.20
pandas>=1.3
astropy>=5.0
matplotlib>=3.5
pyyaml>=5.4
scipy>=1.7
